// packages/app/src/components/settings/types.ts
export interface Settings {
  darkMode: boolean;
  containerWidth: number;
  sortEnumsByValue: boolean;
  hideInheritedMembers: boolean;
  defaultExpandAccordions: boolean;
  hideFlagsInLists: boolean;
  searchOmitChildren: boolean;
  searchCheckCase: boolean;
  searchCheckFullPath: boolean;
  searchListCount: number;
}